# consensus/ethash: implement faster difficulty calculators