--- a/consensus/ethash/consensus_test.go
+++ b/consensus/ethash/consensus_test.go
@@ -17,6 +17,7 @@
 package ethash
 
 import (
+	"encoding/binary"
 	"encoding/json"
 	"math/big"
 	"math/rand"
@@ -87,44 +88,33 @@ func TestCalcDifficulty(t *testing.T) {
 	}
 }
 
-func TestDifficultyCalculator(t *testing.T) {
-	x := makeDifficultyCalculator(big.NewInt(1000000))
-	y := MakeDifficultyCalculatorU256(big.NewInt(1000000))
-	rand.Seed(2)
-	for i := 0; i < 50000; i++ {
-		// 1 to 300 seconds diff
-		var timeDelta = uint64(1 + rand.Uint32()%300)
-		var difficulty = make([]byte, 10)
-		rand.Read(difficulty)
-		h := &types.Header{
-			UncleHash:  types.EmptyUncleHash,
-			Difficulty: new(big.Int).SetBytes(difficulty),
-			Number:     new(big.Int).SetUint64(rand.Uint64() % 50_000_000),
-			Time:       rand.Uint64() - timeDelta,
-		}
-		exp := x(h.Time+timeDelta, h)
-		got := y(h.Time+timeDelta, h)
-		if exp.BitLen() > 256 {
-			continue
-		}
-		if exp.Cmp(got) != 0 {
-			t.Fatalf("test %d: error, got \n%x\n, exp \n%x\nHeader: \n%v\n", i, got, exp, h)
-		}
-	}
+func randSlice(min, max uint32) []byte {
+	var b = make([]byte, 4)
+	rand.Read(b)
+	a := binary.LittleEndian.Uint32(b)
+	size := min + a%(max-min)
+	out := make([]byte, size)
+	rand.Read(out)
+	return out
 }
 
-func TestDifficultyCalculatorFrontier(t *testing.T) {
+func TestDifficultyCalculators(t *testing.T) {
 	rand.Seed(2)
-	for i := 0; i < 50000; i++ {
+	for i := 0; i < 5000; i++ {
 		// 1 to 300 seconds diff
-		var timeDelta = uint64(1 + rand.Uint32()%300)
-		var difficulty = make([]byte, 10)
-		rand.Read(difficulty)
+		var timeDelta = uint64(1 + rand.Uint32()%3000)
+		diffBig := big.NewInt(0).SetBytes(randSlice(2, 10))
+		if diffBig.Cmp(params.MinimumDifficulty) < 0 {
+			diffBig.Set(params.MinimumDifficulty)
+		}
+		//rand.Read(difficulty)
 		header := &types.Header{
-			UncleHash:  types.EmptyUncleHash,
-			Difficulty: new(big.Int).SetBytes(difficulty),
+			Difficulty: diffBig,
 			Number:     new(big.Int).SetUint64(rand.Uint64() % 50_000_000),
 			Time:       rand.Uint64() - timeDelta,
+		}
+		if rand.Uint32()&1 == 0 {
+			header.UncleHash = types.EmptyUncleHash
 		}
 		bombDelay := new(big.Int).SetUint64(rand.Uint64() % 50_000_000)
 		for i, pair := range []struct {