@holiman 2020-12-08
Hm interesting, this doesn't actually fail on `bigEndian` systems. 
```
[user@work ethash]$ GOARCH=ppc64 go test -c
[user@work ethash]$ qemu-ppc64-static ethash.test -test.v
```
All good then!